gesture: fix double tap doesn't always work in suspend
Fix bug where sometimes on suspend tap for battery would never work,
but open a resume and suspend again it would work fine. Problem is
that if suspended when history_index is 1, (1 in 40 chance), then
we would never run the detection algorithm, because the check for
if the history buffer has been initialized is incorrect.

This also fixes the algorithm so that on suspend, it requires the
full sensor history buffer  be filled up again before starting to
detect the double tap.

BUG=none
BRANCH=samus
TEST=go in to suspend when history_index is 1 and verify that tap
for battery works.

Change-Id: I2ba4ab2c807ec6ac1885a4829efedac3c83b32f1
Signed-off-by: Alec Berg <alecaberg@chromium.org>