driver: bmi160: Add command to list last 128 SPI commands.
Store in a circular queue read and write accesses.
Retrieve these access with 'spilastcmd'

BRANCH=smaug
BUG=chrome-os-partner:47522
TEST=Check spilastcmd output makes sense.
Note that given there is a lot of printf, the BMI160 FIFO queue
is not service fast enough and overflow. This is not a problem if
we issue that command while the sensors are stuck.

Change-Id: I87d6515b115f8e22361bb1e2e906b1a75be1d601
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>