Cause "make buildall -j" to yell about failures.
When building with -j, it's easy to miss errors. If you don't
have your shell configured to warn you about nonzero exit, you
may not even notice that "make buildall -j" failed. To make it
more obvious, we'll do one level of recursion for that target.
That will ensure that the last line of output is always a
pass/fail message.

BUG=none
BRANCH=none
TEST=make buildall -j   (with and without errors)

Before this CL, a successful "make buildall -j" ends with this:

  MV      ec.bin
  OBJCOPY ec.hex
  LD      RW/ec.RW.elf
  NM      RO/ec.RO.smap
  OBJCOPY RO/ec.RO.flat
  NM      RW/ec.RW.smap
  OBJCOPY RW/ec.RW.flat
  CAT     ec.obj
  OBJCOPY ec.bin
  COPY_RW ec.bin
  MV      ec.bin
  buildall completed successfully!
  (cr) ((fa7baa6...))  ~/trunk/src/platform/ec $

while a failing one looks like this:

  MV      ec.bin
  OBJCOPY ec.hex
  LD      RW/ec.RW.elf
  NM      RO/ec.RO.smap
  OBJCOPY RO/ec.RO.flat
  NM      RW/ec.RW.smap
  OBJCOPY RW/ec.RW.flat
  CAT     ec.obj
  OBJCOPY ec.bin
  COPY_RW ec.bin
  MV      ec.bin
  (cr) ((fa7baa6...))  ~/trunk/src/platform/ec $

Did you see the difference? I suspect some people miss it.

With this CL, a failing "make buildall -j" looks like this:

    NM      RW/ec.RW.smap
    OBJCOPY RW/ec.RW.flat
    CAT     ec.obj
    OBJCOPY ec.bin
    NM      RW/chip/mec1322/lfw/ec_lfw-lfw.smap
    OBJCOPY RW/chip/mec1322/lfw/ec_lfw-lfw.flat
    COPY_RW ec.bin
    MV      ec.bin
    COPY_RW ec.bin
    MV      ec.bin
    OBJCOPY ec.bin
    COPY_RW ec.bin
    MV      ec.bin
  make[1]: Leaving directory '/mnt/host/source/src/platform/ec'
  Makefile.rules:93: recipe for target 'buildall' failed
  make: *** [buildall] Error 2
  (cr) (stopit)  ~/trunk/src/platform/ec $

Change-Id: Id9b47d2869f61e8e3e44b3c618399ca9223f0a71
Signed-off-by: Bill Richardson <wfrichar@chromium.org>