Update patch set 1
Patch Set 1:

As for the other areas such as hibernate that are implicitly using gpio_set_flags_by_mask() for setting mux mode to gpio, we can pair it with a call to gpio_set_alternate_function(port, mask, func) with func set to -1. That makes it clear that we are setting mux mode to gpio. This is already in uart_enter_dsleep(void) in chip/mec1322/uart.c

Patch-set: 1