cortex-m: Create .bss.slow section.
For those ECs with external storage and a little firmware (LFW) loader,
we can reuse the space that was occupied by the LFW.  Once we're running
in main(), we no longer need the loader in RAM.  Therfore, we can use
this space for more data.  Currently, this is 3k on mec1322.  Lastly,
this region for mec1322 is code optimized RAM, so ideally infrequently
used data should go here.

BUG=chrome-os-partner:46056
BUG=chrome-os-partner:46063
BRANCH=None
TEST=Build GLaDOS, check output memory map for LDR_REGION and .bss.slow
section.
TEST=make -j buildall tests

CQ-DEPEND=CL:305254

Change-Id: Ia74a0df275e057b029749a3a422eb3fecb047153
Signed-off-by: Aseda Aboagye <aaboagye@google.com>