Update patch set 2
Patch Set 2:

Yes. Problem is happen only bootup time, this is enough to go into boot up. Nobody set the default value(means enable it again so that we used all the past time) for this. So this undoing will be the only code who cares these two bit i think

Patch-set: 2