Update patch set 1
Patch Set 1:

> This is a new review because you modified the 'Change-Id' field. If
 > you modify 'Change-Id' to be Ic0ab3e57679fc7ea98a7b73527ce2276e706db1d
 > then it will show up as revision to your previous change, and then
 > you can abandon this review.

Well, my intention is to abandon the old review and keep this one. Because the old one's modifications was base on kunimitsu of the version on 10/13 and this one is base on 10/15 which added four new Merged on 10/14 & 10/15.
But if you prefer modifying the Change-Id, I will abandon this one.

Patch-set: 1
Reviewer: Gerrit User 1137259 <1137259@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0