Update patch set 2
Patch Set 2:

(2 comments)

bq24773.c doesn't belong to Celes, and this CL is starting to look more like a workaround to fix a particular issue on your board. Let's follow the recommendation -- set / clear the option bits based on battery attachment. That's a change that seems reasonable for all boards that use this charger IC.

Patch-set: 2
Label: Code-Review=0