Update patch set 3
Patch Set 3: Verified+1

Alright, so here's the latest patch which incorporates Vincent's feedback. Regarding the exception encountered in the loader, I have managed to narrow it down. Firstly, in order to reproduce the issue, one has to sysjump very rapidly between RO and RW (within about 1 second). Having done that, maybe once in 20 attempts the exception is encountered.

The exception is a forced hard fault which is a result of jumping to a bogus address (either all 0's or all 1's). It seems that sometimes when we load either the RO or RW image from the external SPI it will load all 0's or all 1's. When we fetch the reset vector from the image, we end up grabbing the bogus address of all 0's or 1's.

Now, this currently happens on ToT without my changes applied so it's pre-existing. But, I'm not sure as to the reason why it's occurring. The addresses that we're loading from external SPI are correct and they're hardcoded into the image via macro definitions. It's just that sometimes when you do this really quickly we get a bad read or something.

I added error checking around the SPI transactions and no errors are reported in the SPI transaction even when this exception occurs.

Anyways, since it's pre-existing and not related to this change, I decided to upload this patchset. It only manifests itself when sysjumping very rapidly, but I would like to know why this happens.

Patch-set: 3
Label: Verified=+1