Update patch set 1
Patch Set 1:

Hi Rong,

I disabled Turbo mode and it didn't help. Still see discharging. Checked with Behzad, our hardware engineer, he said "there is a diode in parallel with BAGTE which always would be on if the battery voltage is higher than VBATA. It will discharge again" and also we don't want to disable Turbo functionality.
when the battery is 100%, the requested charging voltage is 8700mV, not 0. It's charging/battery logic in software force it's to 0 while current to 0. Other manufactures' chargers have no problem with this, but seems it's not working well with Intersil charger. If we set the voltage as battery indicates, then it's Ok here.

Patch-set: 1