Update patch set 1
Patch Set 1:

> just a generic comment to start with :
 > are you re-implementing what the 'cmd' python module is already
 > doing ?
 > (as in https://docs.python.org/2/library/cmd.html )


Oh wow, I didn't know about that module; it seems really helpful! I guess in a way, I am re-implementing that. Although, I'm not sure if it'll work with PTYs. I need to be able to handle every keystroke for editing instead of just a readline. I'll check it out and see if it'll work with PTYs. Thanks for bringing it to my attention.

Patch-set: 1