Update patch set 2
Patch Set 2: Verified+1

(4 comments)

So, I tried to use the 'cmd' module in the standard library, but it didn't work with the PTYs. First, it was defaulting to stdin. I got around that but then it seemed like it was trying to readline(). Echo-ing of the characters didn't work either. The module seems helpful if I wasn't using PTYs for the interface.

Fixed a bug with this patchset where the console would lockup if you typed more than the line limit.

Patch-set: 2
Label: Verified=+1