Update patch set 1
Patch Set 1:

(1 comment)

Li1, setting voltage=current=0 seems to be for legacy reasons with old chargers. Changing it in the smart battery driver would potentially break older chargers, but also would not fix the problem here. There are many places in the code we set voltage=current=0 in the charge state machine and all of these would have problems with the intersil charger.

Patch-set: 1