Update patch set 2
Patch Set 2:

> Fine but requires a large amount of testing to ensure that the
 > clock is always up before the first message and switching off the
 > clock is not putting us into a weird state if it happens in the
 > middle of a message

We are switching the clock of in deep sleep part of idle(). It won't happen until 500us after AP is out of S0 (CONFIG_LOW_POWER_S0 not set): so there won't be any i2c message pending at that time.


Looking at transition between S3 and S0, idlestats indicates we are at least 50us from the next deadline.

Note that due to a previous bug, time was set for fi2cclk at 48Mhz even when HSI clock was used.

I notice that we do not alter the bit 5 of RCC_CFGR3 in older version of the EC firmware, so once RW set it to 1 (to use SYSCLK/48Mhz), RO will use that clock as well. Given we are using 48Mhz i2c timing parameter, we still can communicate with the host.

Patch-set: 2
Label: Code-Review=0