make: decouple rw and ro object file sets
When building EC image, in the majority of cases the RW and RO images
are built from exactly the same set of object files, and the RO set of
objects is used as a template to derive the RW set of objects.

This is not necessarily correct in all cases, let's just create an
abstract set of object files and use it to derive the sets for RO and
RW as appropriate.

BRANCH=None
BUG=chrome-os-partner:43025
TEST=tested as follows:
  - changed the Makefile to sort all object files in a single list
    (instead of sorting them by directory, with the directory list
    unchanged). Built all targets, saved all .smap files. Then applied
    this change and again built all targets. Compare all smap files,
    there were no differences.

Change-Id: Ie563aca62028cae9e16f067ba20b5e2930355cf5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>