Update patch set 1
Patch Set 1:

First upload for review. Same as email except I added a link to the servo page.

Need Willis to finish the block diagram and provide pdf schematic. So first time round please focus on the text.

I put this in the board directory since that seems the natural place for the schematic. Other suggestion was in doc/

Patch-set: 1
Label: Code-Review=0