common.h: Create __bss_slow tag.
Some ECs such as the MEC1322 have a data RAM optimized region as well as
a code RAM optimized region.  We discovered that we could save quite a
bit more space by reusing the a portion of the code RAM region as an
additional .bss section.  However, this region resides in the code RAM
region.  If on the same cycle the processor fetches an instruction and
does a load or store to this code RAM region, the data access will be
delayed by one cycle.  Hence, the naming of ``.bss.slow" section.

For boards which do not define CONFIG_REUSE_LOADER_WITH_BSS_SLOW, all
objects bearing this tag will be simply appended to the existing .bss
section.

BUG=chrome-os-partner:46056
BUG=chrome-os-partner:46063
BRANCH=None
TEST=make -j buildall tests

Change-Id: I126fbeee5255732a6dd6fea1d4557fc2b2c62c96
Signed-off-by: Aseda Aboagye <aaboagye@google.com>