Update patch set 7
Patch Set 7:

We use HOOK_CHIPSET_SUSPEND in many places throughout the code. Is it fair to say that we'll mostly want to take the same actions going to S0ix as we do going to S3? If so, we probably don't want to add a new hook. Can we trigger HOOK_CHIPSET_SUSPEND when we go to S0ix, and then for the rare cases that we do need to take different actions for S0ix vs S3, call a function to check the actual power state we're going to from the hook routine?

Patch-set: 7
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0