Cr50: Update to the "final" FPGA image(s)
This provides support for three FPGA images:

  20151104_011218 - full crypto, no USB
  20151104_041733 - tiny crypto, full USB
  20151104_065845 - full crypto, full USB (only for hard-to-get boards)

This also adds a check for the USB features at run time, so that it's
safe to build with CONFIG_USB and run on a no-USB FPGA image.

BRANCH=none
BUG=chrome-os-partner:43791
CQ-DEPEND=CL:310978
TEST=make buildall

I also built the Cr50 firmware and tried it on both the crypto
and USB FPGA images. Both worked as expected.

Change-Id: Ia8a064758f71f86771729437ae3e81226fd55789
Signed-off-by: Bill Richardson <wfrichar@chromium.org>