Cr50: Disable RO image skeleton by default
A previous commit caused ToT to use a not-yet-working bootloader.
This disables that bootloader by default so that the rest of us
can continue to work. ;-)

To re-enable, just build with VADIM=1 on the command line.

BRANCH=None
BUG=chrome-os-partner:43025, chromium:551151
TEST=make buildall

The normal "make buildall" installs on the Cr50 and works as
before. To build the RO skeleton instead, just say

  VADIM=1 make BOARD=cr50

or

  make buildall VADIM=1

You may have to run "make BOARD=cr50 clean" first when switching,
of course.

Change-Id: Ie433b07860cb1b04c12b2609c6fa39025fc0e515
Signed-off-by: Bill Richardson <wfrichar@chromium.org>