cr50: util: signer: make signer header definition usable on the target
The signer running on the build host and the loader running on the
target must be in sync as of the structure of the signature header.

To be able to use the same definition in both programs, remove the
system includes from the .h file which needs to be shared. Rearrange
includes in image.cc to follow the coding conventions.

BRANCH=none
BUG=chrome-os-partner:43025
TEST=with the rest of the patches applied the code successfully boots,
     which indicates that the signer is in fact working properly. (See
     the top patch for testing details).

Change-Id: I6bc9c57ebea55ac256fcdac8338c5566f16b6371
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>