GLaDOS: Kunimitsu: Enable link-time optimization.
Turn on LTO for GLaDOS and Kunimitsu.  This saves about 5k from the
image on GLaDOS.  Also, LTO is disabled for the loader since it actually
causes it to bloat in size for some reason.

BUG=chrome-os-partner:46063
BRANCH=None
TEST=Build and flash on GLaDOS with charger inserted.  Verify that EC
boot is successful.  sysjump to RW and verify that the jump is
successful.
TEST=make -j buildall tests

Change-Id: I9892edfc724f290acaf6cceba181c177702d63bf
Signed-off-by: Aseda Aboagye <aaboagye@google.com>