Update patch set 1
Patch Set 1:

> Keyboard scanning DOES run very quickly.
 > 
 > We scan the entire keyboard every 1 ms, with only 10-20 us between
 > rows.


Okay, so I'm trying to find a cut-off period. This seems like a borderline case to me right now. Here's what I'm thinking. The 48MHz clock is ~21ns per cycle. Assuming worst case scenario of 5 cycles per instruction (Load to PC + pipeline refill). That's about 0.105us per instruction. If we allow a factor of 10x for "room", that's 1.05us.

So if we're accessing a variable in .bss.slow every 10us, it would now take 10.021us. Is that so bad and does that make sense?

If we set the limit at no back to back accesses within 5us, that seems reasonable to me*. That's still a factor of 50x for "room".

*Given that my above reasoning is sound and I'm not missing anything which I totally could be. :)

Patch-set: 1