Update patch set 8
Patch Set 8:

(2 comments)

> a. GPIO reading inside interrupt returns always 0, which means there is interrupt laterncy and reading happens between 4 ~ 5.

So, interrupt latency is always greater than 1ms? I guess it auto-solves our debounce problem. 

> b. instead reading GPIO reading at '3', 'in_signals' at '1' was ?used inside interrupt temporarily.

Why make this change? Because the interrupt never saw the GPIO in high state? But that seems totally fine, and consistent with how we want to handle the signal.

> c. to ignore interrupt at '4' by disabling interrupt between '2' ~ '4'.
> d. if SLP_S0 is de-asserted(by power button/lid open) at 'A', it will be ignored because interrupt was diabled before.

We'll need to fix that. Why did you decide to disable the s0ix interrupt to begin with?

Patch-set: 8
Label: Code-Review=0