Update patch set 8
Patch Set 8:

a. GPIO reading inside interrupt returns always 0, which means there is interrupt laterncy and reading happens between 4 ~ 5.

So, interrupt latency is always greater than 1ms? I guess it auto-solves our debounce problem.

--> ISR still starts with delay but not fast enough, right after rising edge.
    SLP_S0(GPIO141) shares same NVIC with UART in interrupt group(8).  I think UART interrupts is one of the reason.

    b. instead reading GPIO reading at '3', 'in_signals' at '1' was used inside interrupt temporarily.

Why make this change? Because the interrupt never saw the GPIO in high state? But that seems totally fine, and consistent with how we want to handle the signal.
--> when interrupt happens, if levels before("1") and after("5") interrupt are different, we can determine this interrupt is valid.

    c. to ignore interrupt at '4' by disabling interrupt between '2' ~ '4'.
    d. if SLP_S0 is de-asserted(by power button/lid open) at 'A', it will be ignored because interrupt was diabled before.

We'll need to fix that. Why did you decide to disable the s0ix interrupt to begin with?
---> by diabling interrupt temporarily at the begin, interrupt at "4" can be ignored which is what we want.
     SLP_S0 interrupt will be re-enbled right after "5", at the beginning of referred function.
     Cases at 'A' and 'B' were already considerd and this will cover/fix asynchronous assertion of SLP_S0 by lid-open/powerbutton.

Patch-set: 8