Update patch set 9
Patch Set 9:

I don't understand why we need to ignore the interrupt on the falling edge of the GPIO right after the 1ms pulse. The interrupt will see the GPIO state as low and correctly set slp_s0_debounced = 0. What's the harm?

Also I don't think we should be sampling the power_get_signals() copy of the power signals. Sampling the actual GPIO should be sufficient to debounce it. Depending on the shadow copy adds more possible races and adds needless complexity.

Patch-set: 9