DO NOT COMMIT: sign discovery-stm32f072 as usbpd1
This signs the discovery-stm32f072 using the (deprecated) usbpd1
format, and jumps from RO to RW if the signature is valid.

The keypair was generated with

  openssl genrsa -F4 -out board/discovery-stm32f072/dev_key.pem 2048

BUG=none
BRANCH=none
TEST=manual

You'll want to prepare your workstation to see the USB consoles
exported by the firmware. Do this once, outside the chroot:

  cd extra/usb_serial
  sudo cp 51-google-serial.rules /etc/udev/rules.d
  sudo udevadm trigger
  make
  sudo insmod ./raiden.ko

Then you can open the console with

  minicom -D /dev/google/discovery-stm32f072*/serial/Shell

To build and install the image, do this inside the chroot:

  cd ../platform/ec
  make BOARD=discovery-stm32f072 -j
  make -n -C board/discovery-stm32f072 flash

On the console, you can run "sysinfo" to see whether it's running
the RO or RW image. Of course, since the console only exists while
the board is running, it goes away during a reboot and only comes
if everything worked.

Change-Id: I37a2a1c19d92cdaf9985fad0c6faceec2df83f42
Signed-off-by: Bill Richardson <wfrichar@chromium.org>