Update patch set 1
Patch Set 1:

I think I understand what this change does.

Interrupt event is set if either RX or TX event is set.

When you process RX, that clears the RX event, but the TX event may still be set if there is TX data to process.

But if more RX data comes in during TX event processing, that will set the RX event again before the TX event clears.

So the interrupt event = (RX|TX) never deasserts.

So there's not another rising edge of the interrupt event.

So when the ISR exits, another interrupt isn't generated even though there's been more received data.

So, instead, copy out the current IER values, zero IER, and then replace those values.  That will cause the interrupt event to go to 0, and then it will go back to 1 if RX or TX event became asserted after processing in the ISR.

And this works even in the following case:

1. RX=TX=1
2. Process RX (RX=0, TX=1)
3. Process TX (RX=TX=0)
4. Read IER (=0, no new data has come in yet)
5. New data comes in after the read instruction.  Sets RX=1.
6. Set IER=0.  RX=TX=0.
7. New interrupt fires.
8. RX=TX=0 still
9. But we still call uart_process_input(), which uses LSR register rather than ISR to determine if there's data to read.

If step 9 used RX=1 to decide whether to call uart_process_input() it would be a problem.  But since it doesn't, I think it's ok.

And the new interrupt will still fire, because there is a rising edge of the interrupt event in steps 4->5 of my case above.

Am I understanding that correctly?

Patch-set: 1