Update patch set 1
Patch Set 1:

So, what we're claiming is effectively:

Even if *everything* were in __bss_slow, with the 48 MHz clock, we'd get equivalent performance as if we were running at 24 MHz.

All this code runs fine on LM4 at 16 MHz.  So it should still run fine with its data in __bss_slow.

I guess that's reasonable.

Patch-set: 1