Update patch set 3
Patch Set 3:

Yes the MEC has 128kB of ram, and yes 96kB is optimized for code and the remaining 32kB is optimized for data.
You can however use the code ram for data and the data ram for code.
The 2 regions of memory are contiguous, i.e. the address of the last byte of code optimized memory is one less than the address of the first byte of data optimized memory.

In addition there is a purely artificial CONFIG_RO_SIZE region. Into this the linker is instructed to put the code and the RO data. This is mapped starting 3kB into the code optimized ram, and (since it is now 104kB in size) the first 93kB fills the rest of the code ram and then 11kB fills the first part of the data ram. The figures last time I looked showed that there was about 75kB of code, and this is put in the CONFIG_RO_SIZE region first, so what we actually have is (starting from the bottom of code ram)
------------- start of (code) ram
3kB for the little firmware
75kB for the main code image
12kB part of the RO data
---------------- end of code ram, start of data ram
11kB rest of RO data
          ------ end of artificial CONFIG_RO_SIZE section.
21kB for RW data
---------------- end of (data) ram

I have proposed scrapping the CONFIG_RO_SIZE section, and just putting everything into one region. This would avoid the need to keep tweaking the boundary every time we scrape together a few more bytes of savings, and also would eliminate the wasted bytes between the two regions. In a conventional Unix type program having the RO data and the code together makes sense, all of it can be shared between running instances of the program and it can be protected, and there is usually only one type of ram. It makes a lot less sense in this embedded case.

The background to this change and another is that we have run out of space in the CONFIG_RO_SIZE region, so are trying to remove things from the RW region.

Your other question about the RAM usage. As is common for embedded designed there is no heap (so no malloc/free) and no task creation (i.e. nothing like 'fork' or 'pthread_create'). All memory usage is designed up front, the stack sizes are fixed (i.e. they will not grow in the way that the stack grows on a unix system on demand). So the ram usage can just be measured at link time. It does require you to think in a slightly different way, there are fixed limits everywhere and you need to worry about them. The EC does have a few buffers that a task (think thread or process) can borrow and then return so this gives you some of the flexibility that you would typically use malloc/free for.

Hope this answers your questions.

Patch-set: 3
Reviewer: Gerrit User 1110034 <1110034@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0